+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.1.7
+
+### Patch Changes
+
+-   59a24b3: FF & FF writer fix enableNPMWorkspaces option for CAP
+
+## 2.1.6
+
+### Patch Changes
+
+-   1baff46: Add error handling to fiori elements writer while generating annotations
+
 ## 2.1.5
 
 ### Patch Changes